feat(model): bf16 default precision end to end (ROADMAP F5)
Checkpoint torch_dtype now drives ModelConfig.dtype (bf16 when undeclared,
fp32 downgraded); the loader materialises parameters and widens FP8 weights
in that dtype instead of blanket fp16. Triton kernels (flash attention,
flash decoding, w4a16/w8a16/w8a8/fp8 linear) accept bf16 activations --
tl.dot operands follow the activation dtype, accumulation stays fp32.
fp16 checkpoints keep fp16 parameters. Golden baselines regenerated in
bf16 (Qwen2.5-0.5B, Qwen3-0.6B, Qwen3-0.6B-FP8).

+++ lite_llama/executor/loader.py
@@ -27,9 +27,10 @@ from .weight_utils import hf_weights_iterator
 
 logger = get_logger(__name__)
 
-#: Every lite_llama parameter is fp16: that is what the Triton kernels are
-#: written for, and checkpoints are cast on copy.
-PARAM_DTYPE = torch.float16
+#: Parameters take the checkpoint's own element type (``config.dtype``) — bf16
+#: unless the checkpoint declares otherwise. Kept as a named constant only for
+#: the rare caller that builds a model with no config at hand.
+PARAM_DTYPE = torch.bfloat16
 
 
 @contextlib.contextmanager
@@ -140,12 +141,17 @@ class DefaultModelLoader:
         logger.info("Building %s skeleton on the meta device", model_cls.__name__)
         with init_empty_parameters():
             model = model_cls(config)
-        materialise_parameters(model, device)
+        materialise_parameters(model, device, dtype=config.dtype)
 
         # An already-quantised checkpoint is copied in byte for byte; only an
-        # fp16 model wants the FP8 blocks widened on the way through.
+        # unquantised model wants the FP8 blocks widened on the way through.
         model.load_weights(
-            hf_weights_iterator(checkpoints_dir, device, dequantize_fp8=config.quant is None)
+            hf_weights_iterator(
+                checkpoints_dir,
+                device,
+                dequantize_fp8=config.quant is None,
+                dequant_dtype=config.dtype,
+            )
         )
         if quantization and config.quant is None:
             self._quantize(model, quantization)
